(Checking in some old changes.)
In the two places where we use the host syntax-rules, the keyword in
the patterns is now the keyword being bound, rather than "_" (this is
arguably required by r5rs).

Renamed eval-begin to eval-seq (because it only handles expression
sequences, and does not allow definitions like BEGIN does).

Wrote a private version of MAP, because r5rs clearly requires (in the
formal semantics section) the continuations of combination positions
to be handled a certain way, but the r5rs specification for MAP does
not clearly specify what continuations captured during a MAP do.